refactor(studio): 分离观测收件箱交互脚本
